добавила файлы и перенесла код по примеру
